ExceptionInterpretorTest divided into unit and integration
